字节流新增支持切片类型 改进模型读取方式支持递归读取切片类型 测试带符号int型读写 改进断开连接时conn不支持继续传递问题

+++ io/network/packet/hello_world.go
@@ -23,6 +23,6 @@ type Token struct {
 }
 
 type LoginFail struct {
-	Code    uint16
+	Code    int32
 	Message string
 }
